move computed criteria to getCriteria helper. move formattedResults to getFormattedResults helper. move workflowId to getter.
