+++ client/containers/workflow-list/helpers/index.js
@@ -19,7 +19,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+export { default as getCriteria } from './get-criteria';
 export { default as getFilterModeButtonLabel } from './get-filter-mode-button-label';
+export { default as getFormattedResults } from './get-formatted-results';
 export { default as getMinStartDate } from './get-min-start-date';
 export { default as getState } from './get-state';
 export { default as getStatus } from './get-status';
